requirements.txt
asttokens==2.2.1
attrs==22.1.0
Authlib==1.2.0
azure-core==1.26.1
azure-storage-blob==12.14.1
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
config==0.5.1
cryptography==38.0.4
decorator==5.1.1
executing==1.2.0
falcon==3.1.1
gunicorn==20.1.0
idna==3.4
ipython==8.7.0
isodate==0.6.1
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
msrest==0.7.1
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyrsistent==0.19.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
bcrypt==4.0.1
PyJWT==2.6.0
openai==0.25.0
pytest==7.2.0
